pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "nile_upgrades"
version = "0.0.1"
description = "Nile upgrades plugin"
readme = "README.md"
authors = [
  "Eric Lau <ericglau@outlook.com>"
]
license = "MIT"
# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
click = "^8.0.4"

[tool.poetry.dev-dependencies]
darglint = "^1.5.8"
isort = "^5.7.0"
pyupgrade = "^2.7.4"
black = "^20.8b1"
mypy = "^0.981"
bandit = "^1.7.0"
safety = "^1.10.3"
pytest = "^6.2.1"
pylint = "^2.6.0"
pydocstyle = "^5.1.1"
pre-commit = "^2.9.3"
flake8 = "^3.9.2"

# We need to specify that click commands are Poetry entrypoints of type `nile_plugins`. Do not modify this
cairo-nile = "^0.9.0"
[tool.poetry.plugins."nile_plugins"]
# command name and location <command_name> = <package_method_location>
"deploy_proxy" = "nile_upgrades.deploy_proxy.deploy_proxy"
